[FIX] account_reports: don't fetch lines whose deferred dates are entirely inside the report period
Steps to reproduce:
- Set the Manually and Grouped mode in the deferred settings
- Create invoice with deferred_start_date = 1/1, deferred_end_date = 1/31
- Go to the report
- The invoice is shown

--> It shouldn't since there is nothing to be deferred.

This also fixes another bug where a division by zero would happen

Steps to reproduce:
- Set deferred to manually and grouped in the Settings
- Set deferred computation to Equal per month
- Create invoice with start_deferred_date = 01/30/2023, and end_deferred_date = 01/30/2023
- Go to report and generate entry
- Error division by zero

This is because in the "Equal per Month" computation mode, all months
are considered to have 30 days. We will add 1 day to the end_date which
will then be 01/31 which will be reset to 01/30. Thus `_get_deferred_diff_dates`
will return 0 months which will cause a zero division when computing the `amount_per_month`.

task-id 3544877

Part-of: odoo/enterprise#49758