[IMP] knowledge: permission panel UI
Prior to this commit, there was a spacing issue and also a small
contrast issue with the active state for logged-in users in
the permission panel.
Additionally, the action buttons (such as remove member, leave article)
for hovered-over invited members were not optimally positioned and were
only clickable icons. These have now been upgraded to more user-friendly
buttons.

task-3575300

closes odoo/enterprise#49811

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>