[FIX] pos_preparation_display: combo product kitchen display
Step to reproduce:
- Order two same combo products in pos_restaurent
- Only one product will appear on the kitchen display

Previously the _get_order_changes method was computing the orderline changes based on
a key composed of the product id, the full product name and the note. This was
done considering the fact that the merge orderline was working. Combo products
are not handled by the merge orderline. Because of this, if 2 orderlines of
combo product were the same, only the last orderline would be taken
into account. To prevent this, this commit changes the key to the uuid which
is a unique identifier to an orderline.

closes odoo/enterprise#49893

Related: odoo/odoo#140416
Signed-off-by: David Monnom (moda) <moda@odoo.com>