[IMP] account_reports: display companies in the UI
When having multi company (with branches and tax units) it may sometimes be a
bit difficult for the user to realize what companies are currently used by the
report.

This commit adds under the filters, on top of the report the names of the
companies present in the options dict with the key companies under certain
conditions.

closes odoo/enterprise#49825

Task: 3573675
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>