[FIX] web_enterprise: fix input border color
This commit fixes an issues introduced in Commit[1].
Prior to Commit[1], the `border-color` property of our input was
overridden in Enterprise to a `$o-gray-200` value.

Commit[1] aimed to fix an issue inside list view where the "add item"
button would have a grey background on focus state, by setting the
background to a light gray color.

Unfortunately, this commit didn't take into account the fact that
`$o-form-lightsecondary` is also handling the `border-color` of our input
in the hover state.

By changing it to `$o-gray-100`, it made it barely impossible to notice.

This commit reverts that change, ensuring the `border-color` remains
visible in every state.

Commit[1]: 74bc70a35897274d0bd8d8e53abbdbb48174d8f0

task-3575546

closes odoo/enterprise#50309

X-original-commit: 41494c7fea38f18db07f9908ce75cc7c06ce8429
Signed-off-by: Pierre Paridans (app) <app@odoo.com>
Signed-off-by: Chrysanthe Gomrée (chgo) <chgo@odoo.com>
Co-authored-by: mano-odoo <mano@odoo.com>