[FIX] spreadsheet_edition,_dashboard: reset data
When the field `spreadsheet_data` (or the underlying `spreadsheet_binary_data`
is updated, we should delete the existing revisions because there were
based on the previous data. Loading those revisions on top of the new
data will do random things.

I'm not sure why it was properly managed for spreadsheet dashboards but not
other spreadsheet models. It was also wrong for two reasons:
- spreadsheet revisions can only be deleted by a system user (missing sudo)
- when gargabe collecting revisions (see `_gc_revisions`), all the active
  revisions would also be deleted when resetting the spreadsheet data to
  the last snapshot.

closes odoo/enterprise#49868

Task: 3575942
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>