[FIX] l10n_ch_hr_payroll: fix children field access
Before this commit, a user for a swiss company having
hr rights but no payroll rights wouldn't be able to access employees
because of the `l10n_ch_children` o2m field on `l10n.ch.hr.employee.children`
which is supposed to be only accessed by payroll officers.

This commit adds a group to the field so that only the users with the
correct access rights can see it.

closes odoo/enterprise#49896

Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>