[FIX] l10n_be_reports: add condition to include warnings
Problem:
The user is unable to export or post closing entries due to a traceback
that warnings is of None type. The method, _dynamic_check_lines, tries
to assign an item to warnings which is a dict.
However, since warnings was not initiliazed, it throws a traceback.

Solution:
Initialize warnings as an empty dict when it hits the failed_control condition
This will unblock the user from exporting the tax report or posting the closing entry for the tax report.

Steps to reproduce on Runbot(ALL):
Switch to BE Company CoA
Enable "0% EU S" purchase tax
Create a vendor bill using this tax
Try to post Tax Closing Entry for the period which includes the created vendor bill
Receive technical error
OR
Try to export the Tax Report

opw-3560681

closes odoo/enterprise#50531

X-original-commit: 8a33ee320ac08a19dd1fad1adee7abebeda53535
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>