[FIX] stock_barcode: O-CMD notification
Before this commit, when a barcode starting with "O-CMD." is scanned,
the barcode handler checks if it knows a command and either calls it if
it exists, either displays a notification if not.

The issue with the notification is that some modules can use barcodes
starting with "O-CMD." (`stock_barcode` with "O-CMD.MAIN-MENU" for
example), meaning scanning one of those will display a warning
notification even if the action linked to the barcode is correctly
called.

This commit adds an empty function for each "O-CMD." barcode used by
`stock_barcode` to let `barcodeGenericHandlers` know those commands
exist and should not trigger a warning.

closes odoo/enterprise#49898

Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>