[FIX] project_timesheet_forecast : get correct planned hours
This fix reverts some changes of this commit : https://github.com/odoo/enterprise/commit/2711a3ad30bf306df35042d658c7b62a20fcb96f

Issue:
======
The planned hours in project timesheet forecast is counting the holdiays
as working days too , so the displayed hours are wrong.

Steps to reproduce the issue:
=============================
Activate timesheets and planning in project setting.
Create a product with the following values:
- Type : Service
- Create on order : Project & task
- Plan Services : active and select anything work planner for example
- Unit of Measure / Purchase UoM: Days
Create a sale order with that product and choose 10 quantities and
confirm it.
Click on smart button (8O hours to plan)
Click on new and select a resource and start and end date that have 80
alloctated time and save it.
Go to project , click on the 3 dots on the project related to the sale
order and select Timesheets and Planning analysis
You will get more planned hours than expected , you can groupby day to
see the planned hours for each day.

Origin of the issue:
====================
When creating the table we didn't take into account the calendar of the
employee , so it will always count all days from start to end.

Solution:
=========
I reverted some of the changed of the commit mentioned above which was
responsible for this.

opw-3482252
opw-3563938

closes odoo/enterprise#49901

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>