[FIX] account_reports: deferred fully inside period
Steps to reproduce:
1. Create invoice with date = 1 Jan, deferred start date = 1 Feb, deferred end date = 28 Feb and validate it
2. Open the deferred report on February
3. The invoice does not appear

In January, the amount of the invoice should appear in the 'Later' column, this is the case.
In February, the amount of the invoice should appear in the 'Current' column, this is currently not the case.

This is because the "Exclude if entirely inside the period" domain is
incomplete. It matched all entries whose deferred dates were both inside
the report period and excluded them. However, the invoice date should
also be taken into account so that we do not exclude invoices with an
invoice date outside the report period.

Note: if the invoice date was 1 Feb instead of 1 Jan, then nothing would need
to be displayed/generated since the invoice does not have to be deferred
as it is fully accounted for in February.

Part-of: odoo/enterprise#50096