[FIX] sale_planning, *: display empty line for resources matching search
- impacted modules:
    - sale_planning
    - sale_project_forecast

Steps
=====
- Install module sale_planning
- Create a resource with no shift associated to it
- In the main planning view, search the name of this resource

Issue
=====
No empty line is displayed for this resource.

Cause
=====
The main action of planning is overridden in sale_planning and
sale_project_forecast to change is context. The context key
'planning_expand_resource' was not present in those overridden contexts.

Fix
===
The context key is simply added in the context of the main action in
those two modules.

closes odoo/enterprise#50034

X-original-commit: fedb04f38e02b6845934df582e12f58004c3bf41
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>