[FIX] account_reports: general ledger hierarchy
When having the hierarchy options enabled on the general ledger, some column
with a "string" figure type had the value zero in the cell which was not the
intended behaviour. To correct that we have added a check to change the default
value in case the figure type is a string.

closes odoo/enterprise#50538

Task: 3438422
X-original-commit: 192f549e2fd0263e971a1aecf17bd6dc59efa6e5
Signed-off-by: John Laterre (jol) <jol@odoo.com>