[IMP] l10n_account_customer_statements: remove display config
Adding the display config in this way, with a regular inherit will conflict with
other eventual extension of the partner ledger report.
As this customization targets the standard report itself by just modifying the existing
template, we can remove the python code changing the template and instead simply use
an extension inherit on the template to add the new button.

closes odoo/enterprise#50008

Signed-off-by: William André (wan) <wan@odoo.com>