[FIX] account_reports: filter closing entry when multiple moves
When having tax unit enabled with more than one company, clicking on the
button for the closing entry will redirect to a tree view where the "posted"
filter is put by default. But in this cas we don't want it since all the moves
are in draft. This commit will remove the 'search_default_posted' from the
context when having multiple moves.

closes odoo/enterprise#50048

Task: 3580453
Signed-off-by: John Laterre (jol) <jol@odoo.com>