[FIX] account_{accountant,reports}: division by zero in deferred
Steps to reproduce:
1. Create move with deferred start date = deferred end date = 30 October
2. Validate
3. Error "divison by zero"

Why this happening:
We compute the number of months during which the invoice has to be deferred.
For this, we need to add +1 on the deferred end date to include it.
So, if both the deferred start and date are on the 30 October, then the end_date would become 31 October.
However, we're working with months of 30 days in the case of "Per month" computation, therefore 31 Oct would become again 30 Oct.
So the difference between both dates becomes zero and a divison by zero happens.

Fix:
In such a case, we can simply use nb_months = 1/30 (= 1 day). This is correct
because it follows the same logic as other dates. For instance, if we had
deferred start and end dates on the 20 Oct, we would have a nb_months = 1/30
since we also added +1 day to include the end date, so the difference between
21 Oct and 20 Oct would also be 1 day = 1/30 month.

opw-3547490

closes odoo/enterprise#50096

Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>