[IMP] hr_expense_extract: usability improvements
The method _autosend_for_digitization was called twice when attaching a receipt in an expense. Once due to the widget `attach_document` and another due to an action with the same name. This commit moves the action to the community repo and changes it so it sets the main attachment for the expense instead.

task-3539382

closes odoo/enterprise#50136

Related: odoo/odoo#138387
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>