[FIX] frontdesk: fix favorite widget position
This commit fixes the position of the favorite widget within the kanban
view of Frontdesk.

Prior to this commit, the filter was added after the title, which would
displays it under the title.

To fix the issue, we place it before the title, as we do in Project,
and add some flex classes to handle that layout.

We also removed some `padding` classes to provide a better visual
alignment between the left and right side of our cards.

task-3582084

closes odoo/enterprise#50132

Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>