[IMP] account_reports: composite reports
The goal of this PR is to add a button on the reports next to the filters that
will redirect to the configuration page of the report.
On this page we have changed a placeholder and hide the use_section field in
case the report has lines.

Task: 3569038
Part-of: odoo/enterprise#50142