[FIX] knowledge: fix alignment issues
This commit fixes two alignment issues in the Knowledge sidebar.

> Buttons with the same design having different styling to space the
icon from the text ;

> Buttons not being aligned with the rest of the sidebar element to the
left.

To fix these issues, we adapt our two buttons to make them use the same
styling.

We also have to adapt a bit the overall styling of the sidebar to allow
our buttons to be correctly aligned to the left.

task-3582600

closes odoo/enterprise#50189

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>