[FIX] account_reports: button config
In this commit: https://github.com/odoo/enterprise/pull/50142/commits/1df7c45a2f56a70ae7d1d4aa90f9d0aa1a34cd36
We added a button next to the filters that will redirect to the configuration
page of the report. The problem was that only the cogs was clickable and not the
surrounding background.

closes odoo/enterprise#50248

Signed-off-by: John Laterre (jol) <jol@odoo.com>