[IMP] stock_barcode: split before put in pack
Imagine the following case:
- You have a receipt for 20x a product;
- You want to put 10 product into a first package, then 10 more in
  another package;
- So, you open the receipt in the Barcode App, scan 10 times the
  product's barcode and then put in pack.

What happens here is that the move line is put into a new package but
that's all, so you end up with one line for the product, 10/20 qty and
with the new package as the result package.
Thus, if you scan the product one more time, it will update this
existing and already packaged line's quantity, making impossible to set
the remaining quantity in another package.

To fix that, just before to call the `action_put_in_pack`, the
"reserved" quantity of the move line will be decreased to its `qty_done`
and a new move line will be created for the remaining quantity. Since
the new line won't have any `qty_done`, only the first one will be
packaged.

closes odoo/enterprise#50277

Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>