[FIX] account_reports: remove useless call to hacky l10n_be-specific option key from template
'closing_entry' option key was introduced in l10n_be in order to bypass the display of some options on the pdf stored in attachment of the tax closing entry. It is not needed anymore in account_reports in 17.0, because of the split between pdf and UI templates: the pdf templates' extra_options template only displays information when its options are explicitly set to a value. For example if only posted entries are displayed in the report; it's not written on the pdf ; only the fact of showing draft entries is mentioned when printing. Therefore, the pdf attached to the tax closing moves never shows any of these filters anyway.

closes odoo/enterprise#50292

Signed-off-by: John Laterre (jol) <jol@odoo.com>