[FIX] account_reports: filter closing entry when multiple moves
When having tax unit enabled with more than one company, clicking on the button for the closing entry will redirect to a tree view where the "posted" filter is put by default. But in this cas we don't want it since all the moves are in draft. This commit will remove the 'search_default_posted' from the context when having multiple moves.

closes odoo/enterprise#50445

Task: 3580453
X-original-commit: 70059f37611e2e07f883d0071bd37ab9de310b5e
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>