[FIX] l10n_ec_edi: do not use '_' as variable
Currently, an error occurs as below when cron
'EDI : Perform web services operations' runs.

Error:
`<class 'TypeError'>: "'str' object is not callable" while evaluating
'model._cron_process_documents_web_services(job_count=20)'`

This is because [1] uses '_' as a variable; in Odoo, '_' is used in translation,
and it is a class, so we can't use '_' as a variable.Also, this is causing the
"local variable '_' referenced before assignment" error from [2] because the
'_l10n_ec_cancel_move_edi' method tries to access '_' before [1] when the
above cron runs or when a user requests a cancellation EDI.

This commit fixes the above issue by changing the variable name.

[1]-https://github.com/odoo/enterprise/blob/b1abcb7a5b075ccbb1409a5fb19c6b142bc6cf77/l10n_ec_edi/models/account_edi_format.py#L197
[2]-https://github.com/odoo/enterprise/blob/b1abcb7a5b075ccbb1409a5fb19c6b142bc6cf77/l10n_ec_edi/models/account_edi_format.py#L190

sentry-4608132760

closes odoo/enterprise#50533

X-original-commit: f912a7eba3b3d72dc85056e228f73f1a4d78be02
Signed-off-by: Adrien Widart (awt) <awt@odoo.com>