[FIX] knowledge: ensure promise value for 'isDirty' function
Before this commit:
In several functions, the 'isDirty' function returns a pending promise, causing
the 'if conditions' to always evaluate to true.

Reason:
`The isDirty` has been changed from getter method to async function from PR
https://github.com/odoo/odoo/commit/8723f020c3587a900c811b8cc23f53fe34b98df3

After this commit:
With this fix, the 'isDirty' function now correctly resolves promises, ensuring
that the 'if conditions' are accurately evaluated.

Task-3491577

Part-of: odoo/enterprise#48457