[FIX] documents: avoid partner creation when receiving documents
When receiving a document from an unknown contact, we no longer create a new
partner. However, if the contact is known, we assign the already existing
partner to the document.

We remove that feature because it was creating duplicate partners for documents
sent to accounting which uses VAT number as primary matching criterion. Contacts
created in Documents were messing up with the ones associated with the VAT
numbers.

Issue introduced in the commit: 80db35688eddb1b817c47af2b89a0409a5cc83be.

Task-3579869

closes odoo/enterprise#50810

X-original-commit: 0354c64c6070e376e1c7faca48cddb11aaa56ad9
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Pierre-Yves Dufays (pydu) <pydu@odoo.com>