[FIX] account_reports: allow assigning values to filters of the tax report when coming from the dashboard shortcut
The tax report can be opened from the dashboard, using a shortcut link displayed for each closing move to post. When opened this way, trying to modify a filter on the report never worked.

This happened because the dashboard link uses ignore_session in its related action. With this key, the options are never restored from the session. Since 16.4 and the introduction of sections, the session is however always used to reload the options after any filter assignation. Because of that, loadReportOptions always returned the original options of the report, and modifying a filter manually never had any effect. When coming from the menu item, ignore_session was not used, and the bug didn't occur.

We solve that by changing the meaning of ignore_session and simplifying it: it will now forbid reading options from the session when the report is opened, and nothing more. This means that when opening a report with ignore_session=True, changing some filters, and then opening it again from another action without ignore_session, the options will still be restored. We consider it overall clearer for the user anyway: we always restore the options of the last time the report was opened within the current http session, ignoring the path followed by the user.

closes odoo/enterprise#50421

Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>