[FIX] account_asset: inverse amount for sale asset creation
Before calling `prepare_move_for_asset_depreciation`, the amount needs
to be reversed if we are dealing with a 'sale' asset.
It is currently not the case in the override of the _reverse_moves
method.
This leads to the creation of a depreciation move with a wrong
`depreciation_value` (the one in draft state) and so the
`asset_remaining_value` of the last depreciation move not equal to 0.

Steps to reproduce (clean db with accout_asset installed):
- Go to "Deferred Revenues" and click "New".
- Create a closed sale asset, so set an "Original Value", keep the 5
month default value for "Duration" and set an "Acquisition Date" like
"2023-03-01", so that the duration is over.
Set also a name and the required accounts.
- Validate by clicking "Confirm".  The state should be "CLOSED".
- Access the depreciation moves by clicking the "Posted Entries" smart
button.
- Click on one of the move in order to access its form view, click on
"REVERSE ENTRY" button, then "REVERSE".
- Go back to the sale asset, in the "Revenue Board" tab.
-> The "Cumulative Revenue" (`asset_depreciated_value`) of the last move
isn't equal to 0.

Found this while investigating
opw-3456983

closes odoo/enterprise#50438

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>