[IMP] marketing_automation: add some synchronize tests
Add tests about synchronization and trace update. Purpose is to test a bit more
three sources of synchronization

  * participant creation: creates traces for 'begin' activity;
  * 'sync_participants' method on campaign that creates participants;
  * 'action_update_participants' method on campaign that creates traces for
    participants on activities;

Notably add a test about marketing.trace duplicates generation that happens if
'action_update_participants' is called shortly after syncing participants.
The problematic line is this one
>>> created_activities = campaign.marketing_activity_ids.filtered(
    lambda a: a.create_date >= campaign.last_sync_date
)

Because the 2 dates come from different function (NOW in SQL or datetime.now)
and because the precision is not the same you can have issue if you execute
the action fast. The correct fix is to use env.cr.now (that is exactly for
that type of use case it exists, but it doesn't exist yet in 14.0); hence
adding a delay in the check in 14.0 in next commit.

A test about statistics is also added. In case of multiple participants /
multiple traces per record (manual update, unwanted duplication) we may have
more traces than target records. We consider it is ok, and prefer keeping
statistics highlighting a potential issue compared to trying to guess a 'mean'
trace statistics.

Task-3506681 (MA: Test cleanup)
Prepares Task-2981581 (MA: Fix trace duplication and various issues)

X-original-commit: odoo/enterprise@20e01d71371656c8cb1e615c5dd4b23652fe472a
Part-of: odoo/enterprise#51029
Co-authored-by: Stéphane Debauche <std@odoo.com>
Co-authored-by: Thibault Delavallée <tde@odoo.com>