[FIX] planning: delete false assumptions
`test_recurrency_change_date` was crashing because of false assumptions.
This commit deletes them.

Also, `_repeat_slot` has an argument `stop_datetime`. If it is false, it
is updated inside the loop over self, with the max duration of
generation of the company of the current recurrency, which isn't
especially the same as the next one. With this commit, we name a
variable `recurrency_stop_dateime` which is recomputed on each
iteration.

Finally, that variable was computed by adding (+) the previously mentionned
duration to now, then compared to the occurrence's `start_datetime`. As
long as the former is smaller then the latter, we keep on generating
slots. Yet, the occurrence's `start_datetime` are computed by using
`_add_delta_with_dst`. It leads to inconsistencies when the slot
generation spans over dst. To fix that, we compute
`recurrency_stop_datetime` by using the same function.

task-3449851

X-original-commit: d417c7d7c1a5372909d4c106ae57b0a300987578
Part-of: odoo/enterprise#50613