[FIX] planning: include shifts overlapping recurrency limit
When creating shifts from a recurrency, we determine whether each one
should be open, depending on the `occurring_slots` of the resource.
Before this commit, if a slot started before the end of the recurrence,
and ended after it, we didn't take it into account.
Yet we should, because this shift makes the resource unavailable for its
last shift. The commit fixes it.

task-3449851

closes odoo/enterprise#50613

X-original-commit: a7e22296f904240d078eb2214eb55436b170906f
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>