[FIX] knowledge: add missing template filter in /article dialog
When merging the models `knowledge.article` and `knowledge.article.template`,
we unintentionally forgot to add a condition in the search domain of the
select menu to filter out the templates from the search results. As a
result, the select menu currently lists both articles and templates.
This commit will simply add the missing condition in the search domain
to exclude the templates from the search results.

Steps to reproduce the issue:
1. Open the Knowledge editor
2. Type the /article command
3. Click on the selection menu
=> The select menu lists the articles and the templates.

TO BE: The selection menu should only display the articles and not the
templates

Reference: https://github.com/odoo/enterprise/pull/48691

task-3593373

closes odoo/enterprise#50546

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>