[FIX] marketing_automation: avoid trace duplicates after campaign copy
Copying the 'last_sync_date' field when copying a campaign resulted in the
copied activities to be considered new activities. When an update is performed
on such a campaign this results in traces being duplicated for all activities
present when the campaign performed its first participant synchronization.

Simply setting the last_sync_date as a non-copied field solves the issue.

Task-2981581 (MA: Fix trace duplication and various issues)
opw-2643368

X-original-commit: odoo/enterprise@6dfe76c73f573a599ada641f3d5dbee3b43be7cf
Part-of: odoo/enterprise#51029
Co-authored-by: Stéphane Debauche <std@odoo.com>
Co-authored-by: Thibault Delavallée <tde@odoo.com>