[FIX] l10n_be_reports: backport correct reports from 17.0
In 17.0, we implemented the correct balance sheet and profit and loss
reports for Belgium (commit c76133635bc1542765937116f02bbbd390599106).
In total we have 6 BS versions and 4 P&L versions.

Since the version used in 16.0 is one from 2012 and very outdated, we
will backport the balance sheets and profit and loss report for
companies using the micro/abbreviated scheme, with or without capital.

task-3593997

closes odoo/enterprise#50611

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>