[FIX] sign : do not move 'not editable'sign items
Issue:
======
When you have a sign template which has a sign request it shouldn't be
editable, but accessing it and trying to drag some of sign items will
raise an error because of `helperLines` is undefined.

Steps to reproduce the issue:
=============================
- Install sign
- Go to sign and click on a templte which already have a sign request
  (in demo data : default it's real_estate_listing_agreement)
- Try to drag or resize some sign elements.

Origin of the issue:
====================
We insert a div with absolute position and zIndex 110 so that it covers
the page, however we should also add position:relative in this.root so
that it covers the whole viewer. If you check in your runbot, if you
hover on top of the items in the bottom you still have the cursor as
pointer, while for the first text field, you don't have any cursor
(essentially because the div is covering it).

Note :
======
Adding condition on enableCustom to prevent some useless events
listeners

opw-3579341

closes odoo/enterprise#50867

X-original-commit: c2a32dccf7f833ad14e30eb428da6a9730a23279
Signed-off-by: Leonardo Pavan Rocha <lpr@odoo.com>
Signed-off-by: Mahdi Cheikh Rouhou (macr) <macr@odoo.com>