[FIX] account_reports: Fix how we return the coverage report
The aim of this commit is removing the infinite loader once you
download the accounts coverage report on account.report.
The problem was that we return an ir.actions.act_url with a
target to "self". It means that this url should replace the
current page by the url. This commit changes the target of the
action to "download" according to the documentation
(https://www.odoo.com/documentation/17.0/developer/reference/backend/actions.html#url-actions-ir-actions-act-url)

no task id

closes odoo/enterprise#50665

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>