[FIX] account_bank_statement_import_csv: make sure the rows are sorted
steps to reproduce:
import a csv file where the rows are not sorted by date

before this commit:
balances are not correctly set as the first row and last row are used, assuming
the rows are sorted by date.

after this commit:
raise an error if the rows are not sorted.

opw-3507204

closes odoo/enterprise#50974

X-original-commit: 771ac0040fd3278741e026c3d49f7b7d3c5f5f68
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Nicolas Danhier (nda) <nda@odoo.com>