[FIX] account_bank_statement_import_csv: make sure the rows are sorted
steps to reproduce:
import a csv file where the rows are not sorted by date

before this commit:
balances are not correctly set as the first row and last row are used, assuming
the rows are sorted by date.

after this commit:
raise an error if the rows are not sorted.

opw-3507204

closes odoo/enterprise#50847

X-original-commit: 4cb0a09863e85a276769feb6555e690c3d2fecbd
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Nicolas Danhier (nda) <nda@odoo.com>