[FIX] mrp_workorder: unlink quality check when its move is deleted
Steps to reproduce the bug:
- Create a storable product “C1” tracked by lot
- Create a storable product “P1” with BoM:
    - Quantity: 3
    - component:
       - 1 unit of C1
       - in another line 2 units of C1

- Create a Mo and use the P1 BoM
- Confirm the MO
- Two quality check of “Register consumed materials” type is created because C1 is tracked

- Try to add one line of C1 with 3 units
- The 3 “stock.move” is merged into the first move
- Try to set the qty producing of P1 to 2

Problem:
A traceback is triggered in `_float_check_precision`:
“assert precision_rounding is None or precision_rounding > 0,\
AssertionError: precision_rounding must be positive, got 0.0”

When the moves are merged into the first one, the other moves are
cancelled but their quality checks persist.

So when we change the qty producing, we start the workorder and the
component quantity is updated: https://github.com/odoo/enterprise/blob/1d99f3005f96672fd863541d50e5f81229365c32/mrp_workorder/models/mrp_workorder.py#L100-L102

So, we access the move, but since the move was removed, the rounding
value is set to 0. Consequently, when float_round is called, an error
is raised:
https://github.com/odoo/enterprise/blob/1a6abc731071f39e22693f0c1d54889e3ffdc5ff/mrp_workorder/models/quality.py#L441

https://github.com/odoo/enterprise/blob/1a6abc731071f39e22693f0c1d54889e3ffdc5ff/mrp_workorder/models/quality.py#L445

opw-3484016

closes odoo/enterprise#50752

X-original-commit: fa135f1756298b4008b9e04892d2d14991e4f299
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Djamel Touati (otd) <otd@odoo.com>