[FIX] planning: avoid traceback when selecting roles resources
Steps
=====
- Install module planning
- Go to menu Configuration > Roles
- Click on the resource column of a given role
- In the opened dropdown, click on a resource

Issue
=====
A traceback appears when the resource is clicked.

Cause
=====
The field resource_type is needed for the widget
many2many_avatar_resource and is therefore added as a
relatedFields. However, for field of type 'selection' the selection's
values  have to be added in the field info since the change of the
relational model (see odoo/odoo#114024).

Fix
===
Selection values are added in the field description.

closes odoo/enterprise#50695

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>