[FIX] *_extract: fix "Document not found" error
Since commit a0b24fc, a "Document not found" error could occur in
databases using multiple IAP accounts for the `invoice_ocr` service.

This only occurred when the update of the status of the extraction was
performed through the "Update All Status" cron.

It happened because there was a mismatch between the IAP account
selected in the cron and the IAP account used when the document was
sent.

Now, we force the company set in the context to match the one set on the
document so that the IAP accounts selected will always be consistent.

closes odoo/enterprise#50771

Tickets: 3591259,3592713,3578945,3584204
X-original-commit: bd4bc947266ce4e75987009771d687100b4eb485
Signed-off-by: Florian Daloze (fda) <fda@odoo.com>
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>