[FIX] knowledge: fix embed view toolbar display for templates
This commit slightly adapts the template data to add a br element after the hr
separator, which is necessary to be able to easily click on the embed view
toolbar that appears right next to it.

In addition, we modify the toolbar styling to make it appear "connected" to its
embed view rather than floating a few pixels above it.

closes odoo/enterprise#50724

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>