[FIX] marketing_automation: avoid trace duplicates for new participants
When synchronizing a campaign using 'action_update_participants' traces are
created for activities considered as 'requiring sync'. However traces may
already exist for some participants that have already participated to some
activities. This leads to duplicate traces.

In this commit we now create new traces only for participants that have not
already participated to the activity to update.

Task-2981581 (MA: Fix trace duplication and various issues)
opw-2643368

X-original-commit: odoo/enterprise@3b61a1358ceffd18cb8e9c0669ba549e4a583de7
Part-of: odoo/enterprise#51029
Co-authored-by: Stéphane Debauche <std@odoo.com>
Co-authored-by: Thibault Delavallée <tde@odoo.com>