[FIX] appointment: fix calendar event name
Currently, when booking an appointment, let's say for a tennis court, the
calendar.event name will be setup as follows:
"Tennis Court with Marc Demo".

Which is not very practical for smaller display, as we will have only the first
part of the name remaining: "Tennis Court wi...".

The goal of this commit is to modify it to the following format:
"Marc Demo - Tennis Court Booking".

That way we keep the most important information first, which is the name of the
attendee. This is especially important for the POS Restaurant display.

Task-3599241

closes odoo/enterprise#50817

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>